miApiJWT/requirements.txt
uvicorn[standard]
fastapi
python-jose
python-multipart